Use font server ErrorF/VErrorF instead of private versions
XTrans creates custom ErrorF/VErrorF functions when the including
project doesn't provide them, however the test for that is weak and xtrans
ends up using the private versions for the font server as well.

This means that all xtrans error messages will not be included in the
font server log. It also causes redefinition warnings when building
the font server (which is how this problem was identified).

However, the font server doesn't currently provide a VErrorF function,
so instead of just always relying on the font server to provide these
functions, this patch uses a new TRANS_HAS_ERRORF define to select
whether the project-provided or internal versions will be used. A
patch to the font server that adds VErrorF and defines
TRANS_HAS_ERRORF will be required to fix this bug.

Signed-off-by: Keith Packard <keithp@keithp.com>