Allow partial connection to succeed
Xwayland can optionally be socket-activated by the Wayland compositor,
in which case it would use only the sockets provided by the compositor.

However, that prevents other transport protocols from working, because
when it's given a socket from the Wayland compositor, it would disable
all other connections and rely solely on the given socket.

Change `MakeAllCOTSServerListeners` to allow for partial connections so
that if `partial` is set to a non-zero value by the caller, we do not
bail out in the address is already in use.

That allows to continue trying with other protocols even if the local
connection fails (as this is already handled by the socket from the
Wayland compositor).

Related: https://gitlab.freedesktop.org/xorg/xserver/issues/817
Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>